Fold ReferenceCollectionExtensions.cs into ReferenceCollection.cs
